Add files via upload
In matrix_2x2.py, each ev and pp function and their respective getter and setter functions were redesigned. In addition, float values are now the default type for matrix elements. 

Finally, the files line_graph_indv.py and line_graph_comp.py were combined together into the file line_graph.py.

+++ parameter_estimation.py
@@ -1,6 +1,6 @@
 import sys
 from linear_nudging_alg import LinearNudgingAlg  # lorpy
-
+import numpy as np
 
 def main():
     """
@@ -17,7 +17,7 @@ def main():
         pp_type = sys.argv[2]
         mu_val = int(sys.argv[3])
         relax_time = int(sys.argv[4])
-        bound_val = int(sys.argv[5])
+        bound_val = float(sys.argv[5])
         loop_limit = int(sys.argv[6])
         lin_nudge_alg = LinearNudgingAlg(ev_type, pp_type, mu_val, relax_time, bound_val, loop_limit)
